+++ TestRunner/TestWMMLegendre.cpp
@@ -11,7 +11,7 @@ TEST_CASE("Associated Legendre Polynomial Tests: 0.5")
 {
     using namespace wmm_model;
 
-    constexpr int nMax = 4;
+    constexpr uint16_t nMax = 4;
     constexpr size_t N = (nMax + 1) * (nMax + 2) / 2;
     constexpr std::array<float, N + 1> expected = {1.f, 0.5000f, 0.8660f, -0.1250f, 0.7500f, 0.6495f, -0.4375f, 0.1326f,
                                                    0.7262f, 0.5135f, -0.2891f, -0.4279f, 0.3144f, 0.6793f, 0.4160f};
@@ -24,10 +24,10 @@ TEST_CASE("Associated Legendre Polynomial Tests: 0.5")
     (void)dPcup;
     CHECK(result == true);
 
-    int k = 0;
-    for (int i = 0; i <= nMax; ++i)
+    uint16_t k = 0;
+    for (uint16_t i = 0; i <= nMax; ++i)
     {
-        for (int j = 0; j <= i; ++j)
+        for (uint16_t j = 0; j <= i; ++j)
         {
             char name[128];
             sprintf(name, "n=%d, m=%d", i, j);
@@ -44,7 +44,7 @@ TEST_CASE("Associated Legendre Polynomial Tests MAG_PcupHigh: 0.5")
 {
     using namespace wmm_model;
 
-    constexpr int nMax = 4;
+    constexpr uint16_t nMax = 4;
     constexpr size_t N = (nMax + 1) * (nMax + 2) / 2;
     constexpr std::array<float, N + 1> expected = {1.0000f, 0.5000f, 0.8660f, -0.1250f, 0.7500f, 0.6495f, -0.4375f, 0.1326f,
                                                    0.7262f, 0.5135f, -0.2891f, -0.4279f, 0.3144f, 0.6793f, 0.4160f};
@@ -57,10 +57,10 @@ TEST_CASE("Associated Legendre Polynomial Tests MAG_PcupHigh: 0.5")
     (void)dPcup;
     CHECK(result == true);
 
-    int k = 0;
-    for (int i = 0; i <= nMax; ++i)
+    uint16_t k = 0;
+    for (uint16_t i = 0; i <= nMax; ++i)
     {
-        for (int j = 0; j <= i; ++j)
+        for (uint16_t j = 0; j <= i; ++j)
         {
             char name[128];
             sprintf(name, "n=%d, m=%d", i, j);
@@ -77,7 +77,7 @@ TEST_CASE("Associated Legendre Polynomial Tests: 0.1")
 {
     using namespace wmm_model;
 
-    constexpr int nMax = 4;
+    constexpr uint16_t nMax = 4;
     constexpr size_t N = (nMax + 1) * (nMax + 2) / 2;
     constexpr std::array<float, N + 1> expected = {1.0000f, 0.1000f, 0.9950f, -0.4850f, 0.1723f, 0.8574f, -0.1475f,
                                                    -0.5788f, 0.1917f, 0.7787f, 0.3379f, -0.2305f, -0.5147f, 0.2060f, 0.7248f};
@@ -90,10 +90,10 @@ TEST_CASE("Associated Legendre Polynomial Tests: 0.1")
     (void)dPcup;
     CHECK(result == true);
 
-    int k = 0;
-    for (int i = 0; i <= nMax; ++i)
+    uint16_t k = 0;
+    for (uint16_t i = 0; i <= nMax; ++i)
     {
-        for (int j = 0; j <= i; ++j)
+        for (uint16_t j = 0; j <= i; ++j)
         {
             char name[128];
             sprintf(name, "n=%d, m=%d", i, j);
@@ -110,7 +110,7 @@ TEST_CASE("Associated Legendre Polynomial Tests: 0.9")
 {
     using namespace wmm_model;
 
-    constexpr int nMax = 4;
+    constexpr uint16_t nMax = 4;
     constexpr size_t N = (nMax + 1) * (nMax + 2) / 2;
     constexpr std::array<float, N + 1> expected = {1.000000f, 0.900000f, 0.435890f, 0.715000f, 0.679485f, 0.164545f,
                                                    0.472500f, 0.814127f, 0.331140f, 0.065474f, 0.207938f, 0.828077f,
@@ -124,10 +124,10 @@ TEST_CASE("Associated Legendre Polynomial Tests: 0.9")
     (void)dPcup;
     CHECK(result == true);
 
-    int k = 0;
-    for (int i = 0; i <= nMax; ++i)
+    uint16_t k = 0;
+    for (uint16_t i = 0; i <= nMax; ++i)
     {
-        for (int j = 0; j <= i; ++j)
+        for (uint16_t j = 0; j <= i; ++j)
         {
             char name[128];
             sprintf(name, "n=%d, m=%d", i, j);
@@ -144,7 +144,7 @@ TEST_CASE("Associated Legendre Polynomial Tests: -0.9")
 {
     using namespace wmm_model;
 
-    constexpr int nMax = 4;
+    constexpr uint16_t nMax = 4;
     constexpr size_t N = (nMax + 1) * (nMax + 2) / 2;
     constexpr std::array<float, N + 1> expected = {1.000000f, -0.900000f, 0.435890f, 0.715000f, -0.679485f, 0.164545f,
                                                    -0.472500f, 0.814127f, -0.331140f, 0.065474f, 0.207938f, -0.828077f,
@@ -158,10 +158,10 @@ TEST_CASE("Associated Legendre Polynomial Tests: -0.9")
     (void)dPcup;
     CHECK(result == true);
 
-    int k = 0;
-    for (int i = 0; i <= nMax; ++i)
+    uint16_t k = 0;
+    for (uint16_t i = 0; i <= nMax; ++i)
     {
-        for (int j = 0; j <= i; ++j)
+        for (uint16_t j = 0; j <= i; ++j)
         {
             char name[128];
             sprintf(name, "n=%d, m=%d", i, j);
@@ -178,7 +178,7 @@ TEST_CASE("Associated Legendre Polynomial Tests MAG_PcupHigh: -0.9")
 {
     using namespace wmm_model;
 
-    constexpr int nMax = 4;
+    constexpr uint16_t nMax = 4;
     constexpr size_t N = (nMax + 1) * (nMax + 2) / 2;
     constexpr std::array<float, N + 1> expected = {1.000000f, -0.900000f, 0.435890f, 0.715000f, -0.679485f, 0.164545f,
                                                    -0.472500f, 0.814127f, -0.331140f, 0.065474f, 0.207938f, -0.828077f,
@@ -192,10 +192,10 @@ TEST_CASE("Associated Legendre Polynomial Tests MAG_PcupHigh: -0.9")
     (void)dPcup;
     CHECK(result == true);
 
-    int k = 0;
-    for (int i = 0; i <= nMax; ++i)
+    uint16_t k = 0;
+    for (uint16_t i = 0; i <= nMax; ++i)
     {
-        for (int j = 0; j <= i; ++j)
+        for (uint16_t j = 0; j <= i; ++j)
         {
             char name[128];
             sprintf(name, "n=%d, m=%d", i, j);
@@ -212,7 +212,7 @@ TEST_CASE("Derivative Associated Legendre Polynomial Tests: 0.5")
 {
     using namespace wmm_model;
 
-    constexpr int nMax = 4;
+    constexpr uint16_t nMax = 4;
     constexpr size_t N = (nMax + 1) * (nMax + 2) / 2;
 
     std::array<float, N> Pcuph{};
@@ -228,10 +228,10 @@ TEST_CASE("Derivative Associated Legendre Polynomial Tests: 0.5")
     result = MAG_PcupLow<nMax, N>(Pcup, dPcup, sinf(angle));
     CHECK(result == true);
 
-    int k = 0;
-    for (int i = 0; i <= nMax; ++i)
+    uint16_t k = 0;
+    for (uint16_t i = 0; i <= nMax; ++i)
     {
-        for (int j = 0; j <= i; ++j)
+        for (uint16_t j = 0; j <= i; ++j)
         {
             char name[128];
             sprintf(name, "n=%d, m=%d", i, j);
@@ -248,7 +248,7 @@ TEST_CASE("Derivative Associated Legendre Polynomial Tests: -0.5")
 {
     using namespace wmm_model;
 
-    constexpr int nMax = 4;
+    constexpr uint16_t nMax = 4;
     constexpr size_t N = (nMax + 1) * (nMax + 2) / 2;
 
     std::array<float, N> Pcuph{};
@@ -264,10 +264,10 @@ TEST_CASE("Derivative Associated Legendre Polynomial Tests: -0.5")
     result = MAG_PcupLow<nMax, N>(Pcup, dPcup, sinf(angle));
     CHECK(result == true);
 
-    int k = 0;
-    for (int i = 0; i <= nMax; ++i)
+    uint16_t k = 0;
+    for (uint16_t i = 0; i <= nMax; ++i)
     {
-        for (int j = 0; j <= i; ++j)
+        for (uint16_t j = 0; j <= i; ++j)
         {
             char name[128];
             sprintf(name, "n=%d, m=%d", i, j);
@@ -284,7 +284,7 @@ TEST_CASE("Derivative Associated Legendre Polynomial Tests MAG_PcupHigh: 0.5")
 {
     using namespace wmm_model;
 
-    constexpr int nMax = 4;
+    constexpr uint16_t nMax = 4;
     constexpr size_t N = (nMax + 1) * (nMax + 2) / 2;
 
     std::array<float, N> Pcuph{};
@@ -300,10 +300,10 @@ TEST_CASE("Derivative Associated Legendre Polynomial Tests MAG_PcupHigh: 0.5")
     result = MAG_PcupHigh<nMax, N>(Pcup, dPcup, sinf(angle));
     CHECK(result == true);
 
-    int k = 0;
-    for (int i = 0; i <= nMax; ++i)
+    uint16_t k = 0;
+    for (uint16_t i = 0; i <= nMax; ++i)
     {
-        for (int j = 0; j <= i; ++j)
+        for (uint16_t j = 0; j <= i; ++j)
         {
             char name[128];
             sprintf(name, "n=%d, m=%d", i, j);
